Added a Definition for how to setup the mongo db for testing.
Note: Forgot to note that some of this project began with work by Mohamad Lawand in this  video:

https://www.youtube.com/watch?v=f11nTH3kFHg

I decided to take it a bit further to learn more about how components work in angular.

Setup Add form, but not wired up to save just yet.

What I didn't put in here, is the command to setup the docker instance itself.

(I may decide to add that when I have some time to back track)